+++ app/Livewire/RolesTable.php
@@ -2,10 +2,10 @@
 
 namespace App\Livewire;
 
-use App\Models\Roles;
 use Barryvdh\DomPDF\Facade\Pdf;
 use Illuminate\Support\Carbon;
 use Illuminate\Database\Eloquent\Builder;
+use Masmerise\Toaster\Toastable;
 use PowerComponents\LivewirePowerGrid\Button;
 use PowerComponents\LivewirePowerGrid\Column;
 use PowerComponents\LivewirePowerGrid\Exportable;
@@ -16,33 +16,38 @@ use PowerComponents\LivewirePowerGrid\PowerGrid;
 use PowerComponents\LivewirePowerGrid\PowerGridFields;
 use PowerComponents\LivewirePowerGrid\PowerGridComponent;
 use PowerComponents\LivewirePowerGrid\Traits\WithExport;
+use Spatie\Permission\Models\Role;
 
 final class RolesTable extends PowerGridComponent
 {
     use WithExport;
+    use Toastable;
 
     public function setUp(): array
     {
         $this->showCheckBox();
 
-        return [
-            Exportable::make('export')
-                ->striped()
-                ->type(Exportable::TYPE_XLS, Exportable::TYPE_CSV),
+        $setUp = [
             Header::make()->showSearchInput(),
             Footer::make()
                 ->showPerPage()
                 ->showRecordCount(),
         ];
+
+        if (auth()->user()->can('export')) {
+            $setUp[] =
+                Exportable::make('export')
+                ->striped()
+                ->type(Exportable::TYPE_XLS, Exportable::TYPE_CSV);
+        }
+
+        return $setUp;
     }
 
     public function datasource(): Builder
     {
-        return Roles::query()
-            ->leftJoin('role_permissions', 'roles.id', '=', 'role_permissions.role_id')
-            ->leftJoin('permissions', 'role_permissions.permission_id', '=', 'permissions.id')
-            ->groupBy('roles.id', 'roles.name', 'roles.created_at') // Tambahkan semua kolom yang Anda pilih dari tabel `roles` di sini
-            ->selectRaw('roles.id, roles.name, roles.created_at, GROUP_CONCAT(permissions.name SEPARATOR ", ") as permissions');
+        return Role::query()
+            ->with('permissions');
     }
 
     public function relationSearch(): array
@@ -55,6 +60,9 @@ final class RolesTable extends PowerGridComponent
         return PowerGrid::fields()
             ->add('id')
             ->add('name')
+            ->add('permissions', function (Role $role) {
+                return $role->permissions->pluck('name')->join(', ');
+            })
             ->add('created_at');
     }
 
@@ -65,7 +73,7 @@ final class RolesTable extends PowerGridComponent
             Column::make('Name', 'name')
                 ->sortable()
                 ->searchable(),
-            Column::make('Permissions', 'permissions')
+            Column::make('Permissions', 'permissions', 'permissions')
                 ->sortable(),
             Column::make('Created at', 'created_at')
                 ->sortable()
@@ -80,53 +88,61 @@ final class RolesTable extends PowerGridComponent
         return [];
     }
 
-    #[\Livewire\Attributes\On('edit')]
-    public function edit($rowId): void
+    public function actions(Role $row): array
     {
-        $this->js('alert(' . $rowId . ')');
-    }
+        $actions = [];
 
-    public function actions(\App\Models\Roles $row): array
-    {
-        return [
-            Button::add('edit')
+        // Jika user bisa update
+        if (auth()->user()->can('update')) {
+            $actions[] =
+                Button::add('edit')
                 ->slot('<svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor" class="w-5 h-5">
                 <path stroke-linecap="round" stroke-linejoin="round" d="m16.862 4.487 1.687-1.688a1.875 1.875 0 1 1 2.652 2.652L10.582 16.07a4.5 4.5 0 0 1-1.897 1.13L6 18l.8-2.685a4.5 4.5 0 0 1 1.13-1.897l8.932-8.931Zm0 0L19.5 7.125M18 14v4.75A2.25 2.25 0 0 1 15.75 21H5.25A2.25 2.25 0 0 1 3 18.75V8.25A2.25 2.25 0 0 1 5.25 6H10" />
                 </svg>
                 ')
                 ->id()
                 ->class('pg-btn-white dark:ring-pg-primary-600 dark:border-pg-primary-600 dark:hover:bg-pg-primary-700 dark:ring-offset-pg-primary-800 dark:text-pg-primary-300 dark:bg-pg-primary-700')
-                ->openModal('roles-form', ['rowId' => $row->id]),
-            Button::add('delete')
+                ->openModal('roles-form', ['rowId' => $row->id]);
+        }
+
+        if (auth()->user()->can('delete')) {
+            $actions[] =
+                Button::add('delete')
                 ->slot('<svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor" class="w-5 h-5">
                 <path stroke-linecap="round" stroke-linejoin="round" d="m14.74 9-.346 9m-4.788 0L9.26 9m9.968-3.21c.342.052.682.107 1.022.166m-1.022-.165L18.16 19.673a2.25 2.25 0 0 1-2.244 2.077H8.084a2.25 2.25 0 0 1-2.244-2.077L4.772 5.79m14.456 0a48.108 48.108 0 0 0-3.478-.397m-12 .562c.34-.059.68-.114 1.022-.165m0 0a48.11 48.11 0 0 1 3.478-.397m7.5 0v-.916c0-1.18-.91-2.164-2.09-2.201a51.964 51.964 0 0 0-3.32 0c-1.18.037-2.09 1.022-2.09 2.201v.916m7.5 0a48.667 48.667 0 0 0-7.5 0" />
                 </svg>
                 ')
                 ->id()
                 ->class('pg-btn-white dark:ring-pg-primary-600 dark:border-pg-primary-600 dark:hover:bg-pg-primary-700 dark:ring-offset-pg-primary-800 dark:text-pg-primary-300 dark:bg-pg-primary-700')
-                ->dispatch('delete', ['rowId' => $row->id])
-        ];
+                ->dispatch('delete', ['rowId' => $row->id]);
+        }
+        return $actions;
     }
 
     public function header(): array
     {
-        return [
-            Button::add('add-roles')
+        $header = [];
+        if (auth()->user()->can('create')) {
+            $header[] = Button::add('add-roles')
                 ->slot(__('<svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor" class="w-5 h-5">
                 <path stroke-linecap="round" stroke-linejoin="round" d="M12 4.5v15m7.5-7.5h-15" />
                 </svg>
                 '))
                 ->class('pg-btn-white dark:ring-pg-primary-600 dark:border-pg-primary-600 dark:hover:bg-pg-primary-700 dark:ring-offset-pg-primary-800 dark:text-pg-primary-300 dark:bg-pg-primary-700 w-full')
-                ->openModal('roles-form', []),
-            Button::add('export-pdf')
+                ->openModal('roles-form', []);
+        }
+
+        if (auth()->user()->can('export')) {
+            $header[] =
+                Button::add('export-pdf')
                 ->slot(__('<svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor" class="w-5 h-5">
                 <path stroke-linecap="round" stroke-linejoin="round" d="M6.72 13.829c-.24.03-.48.062-.72.096m.72-.096a42.415 42.415 0 0 1 10.56 0m-10.56 0L6.34 18m10.94-4.171c.24.03.48.062.72.096m-.72-.096L17.66 18m0 0 .229 2.523a1.125 1.125 0 0 1-1.12 1.227H7.231c-.662 0-1.18-.568-1.12-1.227L6.34 18m11.318 0h1.091A2.25 2.25 0 0 0 21 15.75V9.456c0-1.081-.768-2.015-1.837-2.175a48.055 48.055 0 0 0-1.913-.247M6.34 18H5.25A2.25 2.25 0 0 1 3 15.75V9.456c0-1.081.768-2.015 1.837-2.175a48.041 48.041 0 0 1 1.913-.247m10.5 0a48.536 48.536 0 0 0-10.5 0m10.5 0V3.375c0-.621-.504-1.125-1.125-1.125h-8.25c-.621 0-1.125.504-1.125 1.125v3.659M18 10.5h.008v.008H18V10.5Zm-3 0h.008v.008H15V10.5Z" />
                 </svg>
                 '))
                 ->class('pg-btn-white dark:ring-pg-primary-600 dark:border-pg-primary-600 dark:hover:bg-pg-primary-700 dark:ring-offset-pg-primary-800 dark:text-pg-primary-300 dark:bg-pg-primary-700 w-full')
-                ->dispatch('exportPdf', []),
-
-        ];
+                ->dispatch('exportPdf', []);
+        }
+        return $header;
     }
 
     protected function getListeners()
@@ -162,9 +178,9 @@ final class RolesTable extends PowerGridComponent
     // Function to delete data
     public function delete($rowId)
     {
-        $role = Roles::findOrFail($rowId);
-        // Detach all associated permissions
-        $role->permissions()->detach();
+        $role = Role::findOrFail($rowId);
+        $role->revokePermissionTo($role->permissions);
         $role->delete();
+        $this->success('Role telah berhasil dihapus.');
     }
 }
